move all the status code to the correct places so it's still under the process directory, plus return the API Process object rather than the logic object.
